se the justify-content Property in the Tweet Embed
CSS Flexbox: Use the justify-content Property in the Tweet Embed
The last challenge showed an example of the justify-content property. For the tweet embed, this property can be applied to align the items in the .profile-name element.


Add the CSS property justify-content to the header's .profile-name element and set the value to any of the options from the last challenge.