Create Your First CSS Grid
CSS Grid: Create Your First CSS Grid
Turn any HTML element into a grid container by setting its display property to grid. This gives you the ability to use all the other properties associated with CSS Grid.

Note
In CSS Grid, the parent element is referred to as the container and its children are called items.


Change the display of the div with the container class to grid.