Add Rows with grid-template-rows
CSS Grid: Add Rows with grid-template-rows
The grid you created in the last challenge will set the number of rows automatically. To adjust the rows manually, use the grid-template-rows property in the same way you used grid-template-columns in previous challenge.


Add two rows to the grid that are 50px tall each.