Align an Item Vertically using align-self
CSS Grid: Align an Item Vertically using align-self
Just as you can align an item horizontally, there's a way to align an item vertically as well. To do this, you use the align-self property on an item. This property accepts all of the same values as justify-self from the last challenge.


Align the item with the class item3 vertically at the end.