Declare String Variables
Basic JavaScript: Declare String Variables
Previously we have used the code

var myName = "your name";

"your name" is called a string literal. It is a string because it is a series of zero or more characters enclosed in single or double quotes.


Create two new string variables: myFirstName and myLastName and assign them the values of your first and last name, respectively.